Create README.md
The accuracy of statistical analysis is essential for robust and reliable econometric modelling. It is important to assess the presence of statistical phenomena that can affect the validity of financial data, such as foreign exchange rates and economic indicators. In this paper, three important tests are discussed: heteroskedasticity, multicollinearity, and serial autocorrelation, which are important for ensuring the accuracy and integrity of econometric models based on FX data.